CN - Add support for using integer maths instead of floating point for motor mixing. To use the integer math support, define the option MIXER_USES_INTEGER_MATH. There should be no loss of precision using this method. CN - Fix divide with rounding macro CN - Changed DIVIDE_WITH_ROUNDING to function CN - Added integer sine (and cosine) approximation functions. About 15 times quicker than floating point methods, and with negligible loss of precision. CN - Fix some typos. CN - Replaced lrintf with non-standard version. The original rounds .5 values the wrong way (2.5 becomes 2, -2.5 becomes -2), is larger, and is slower. CN - Fix divideWithRounding() function. CN - And fix DIVIDE_WITH_ROUNDING macro to match function name. CN - Added option to use integer approximations for cosf and sinf into board alignment functions. The integer approximation method loses practically no precision, and is around 15 times faster than the floating point method. CN - Altered board rotation so that a rotation context is used. Added in preparation for converting other similar code to use the same technique, doing away with floating point maths at the same time. CN - Added support for converting raw gyro ADC values into degrees/sec x 10. Will be using these values as source to rotation matrix. CN - Work continuing on removal of floating point support in getEstimatedAttitude(). Added i64sqrt() function, which computes the square root of a 64 bit integer. Will probably add a 32 bit version at some stage. CN - Another fix for divide_with_rouunding(). Wasn't working with -ve values. CN - More work on changing things to use fixed point maths rather than floating point. CN - Some changes to integer math rotation code. Won't work well with gyro readings as the resolution is too low. Need to undo and switch back to floating point. CN - Minor changes. Use constant for RAD instead of M_IP/180.0f CN - Remove fixed point changes I made which won't work. CN - Minor changes. Changed vbat calculation to use integer maths. Using DIVIDE_WITH_ROUNDING to do proper rounding of some integers when scaling. CN - More rounding changes. Changed some variables from int16 to int. CN - Remove unused gyro scaling functions. CN - Use integer maths for magnetometer ADC readings. CN - Use integer math for GPS angle calculations CN - Removed compile time option for integer maths in Magnetometer readings. Integer maths is now always used. CN - Removed unnecessary use of floats in sonar transition calculation. CN - Throttle angle correction now uses integer maths. With maximum correction set to 500 and maximum correctin angle set to 45.0 degrees, the maximum difference between floating pont and integer versions is 2. fix typo CN - CLI vartype field changed from unit8_t to enum. Structure is the same size. CN - Sensors values now stored in 32 bit variables rather than 16 bit. Reduces code size. Fix typo in comment. Changed compiler option from gnu99 to c99. Produces smaller code. Image seems to work OK. CN - Add missing atan and sqrt files CN - Add kalman filter and attitude estimation files. Temporarily using estimation from kalman filter to indicate roll and pitch angles. CN - split out some code from the attitude_estimation function. CN - Moved some things around. CN - Remove incorrect comment about integral storage. Couple of minor changes

+++ src/drv_adxl345.c
@@ -36,7 +36,7 @@
 extern uint16_t acc_1G;
 
 static void adxl345Init(sensor_align_e align);
-static void adxl345Read(int16_t *accelData);
+static void adxl345Read(int32_t *accelData);
 
 static bool useFifo = false;
 static sensor_align_e accAlign = CW270_DEG;
@@ -83,10 +83,10 @@ static void adxl345Init(sensor_align_e align)
 
 uint8_t acc_samples = 0;
 
-static void adxl345Read(int16_t *accelData)
+static void adxl345Read(int32_t *accelData)
 {
     uint8_t buf[8];
-    int16_t data[3];
+    int32_t data[3];
 
     if (useFifo) {
         int32_t x = 0;
@@ -109,9 +109,9 @@ static void adxl345Read(int16_t *accelData)
         acc_samples = i;
     } else {
         i2cRead(ADXL345_ADDRESS, ADXL345_DATA_OUT, 6, buf);
-        data[0] = buf[0] + (buf[1] << 8);
-        data[1] = buf[2] + (buf[3] << 8);
-        data[2] = buf[4] + (buf[5] << 8);
+        data[0] = (int16_t)(buf[0] + (buf[1] << 8));
+        data[1] = (int16_t)(buf[2] + (buf[3] << 8));
+        data[2] = (int16_t)(buf[4] + (buf[5] << 8));
     }
 
     alignSensors(data, accelData, accAlign);
